lets just move the deleted file into the recycle bin when appropriate

+++ app/src/main/kotlin/com/simplemobiletools/gallery/models/Medium.kt
@@ -26,7 +26,7 @@ data class Medium(
         @ColumnInfo(name = "size") val size: Long,
         @ColumnInfo(name = "type") val type: Int,
         @ColumnInfo(name = "is_favorite") var isFavorite: Boolean,
-        @ColumnInfo(name = "deleted_ts") var deletedTS: Int) : Serializable, ThumbnailItem() {
+        @ColumnInfo(name = "deleted_ts") var deletedTS: Long) : Serializable, ThumbnailItem() {
 
     companion object {
         private const val serialVersionUID = -6553149366975655L
